move Book conversion to utilities pkg, minor refactor and add notes
