--- src/test/java/org/TeamCodeDefy/googleBooksApi/GoogleBooksConversionTest.java
@@ -1,4 +1,0 @@
-package org.TeamCodeDefy.googleBooksApi;
-
-public class GoogleBooksConversionTest {
-}
